Map JavaScript overhaul
Visitors complained that sometimes the map was blank. I didn't manage to
reproduce this problem, but in an effort to start hunting it I did some
essential overhaul. JQuery was updated. OpenLayers was updated to 2.11.
These libraries were removed from the repository and are now loaded from
cloudflare. Google's map API is not being used any more, and there is no
option to use Google Maps as the base map (only OSM and Greek national
cadastre). Projection 900913 was replaced with the standard equivalent,
3857. In the HTML, JavaScript was moved to the end. Some errors the
browser was showing in the console were also fixed.

It's possible that these changes fix the blank map problem, but I don't
really know.